events-project/requirements.txt
django>=5.1,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-filter>=24.0,<25.0
Pillow>=10.0,<13.0
drf-spectacular>=0.29,<1.0
